requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile base.in
#
asgiref==3.3.1
    # via django
backcall==0.2.0
    # via ipython
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==3.4.7
    # via pyjwt
decorator==5.0.3
    # via ipython
defusedxml==0.7.1
    # via python3-openid
django-allauth==0.44.0
    # via -r base.in
django-extensions==3.1.1
    # via -r base.in
django-filter==2.4.0
    # via -r base.in
django-rest-auth==0.9.5
    # via -r base.in
django==3.1.7
    # via
    #   -r base.in
    #   django-allauth
    #   django-filter
    #   django-rest-auth
    #   djangorestframework
    #   djangorestframework-bulk
    #   drf-yasg
djangorestframework-bulk==0.2.1
    # via -r base.in
djangorestframework==3.12.4
    # via
    #   -r base.in
    #   django-rest-auth
    #   djangorestframework-bulk
    #   drf-yasg
drf-yasg==1.20.0
    # via -r base.in
factory-boy==3.2.0
    # via -r base.in
faker==7.0.1
    # via factory-boy
gunicorn==20.1.0
    # via -r base.in
idna==2.10
    # via requests
inflection==0.5.1
    # via drf-yasg
ipython-genutils==0.2.0
    # via traitlets
ipython==7.22.0
    # via -r base.in
itypes==1.2.0
    # via coreapi
jedi==0.18.0
    # via ipython
jinja2==2.11.3
    # via coreschema
markupsafe==1.1.1
    # via jinja2
oauthlib==3.1.0
    # via requests-oauthlib
packaging==20.9
    # via drf-yasg
parso==0.8.2
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.18
    # via ipython
psycopg2-binary==2.8.6
    # via -r base.in
ptyprocess==0.7.0
    # via pexpect
pycparser==2.20
    # via cffi
pygments==2.8.1
    # via ipython
pyjwt[crypto]==2.0.1
    # via django-allauth
pyparsing==2.4.7
    # via packaging
python-dateutil==2.8.1
    # via faker
python3-openid==3.2.0
    # via django-allauth
pytz==2021.1
    # via django
requests-oauthlib==1.3.0
    # via django-allauth
requests==2.25.1
    # via
    #   coreapi
    #   django-allauth
    #   requests-oauthlib
ruamel.yaml.clib==0.2.2
    # via ruamel.yaml
ruamel.yaml==0.17.2
    # via drf-yasg
six==1.15.0
    # via
    #   django-rest-auth
    #   python-dateutil
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via faker
traitlets==5.0.5
    # via ipython
uritemplate==3.0.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.4
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
